refactor elastic search out of controller
