+++ server/controllers/searchController.ts
@@ -1,14 +1,7 @@
 import express from "express";
-import * as ElasticSearch from "elasticsearch";
 
-import { Config } from "../config";
 import { VuforiaClient as vuforiaClient } from "../vuforiaClient";
-
-interface ElasticSearchSource {
-  success: boolean;
-  id: string;
-  invno: string;
-}
+import { ElasticSearchService } from "../services";
 
 interface VuforiaResponse {
   query_id: string;
@@ -23,40 +16,19 @@ interface VuforiaResponse {
   result_code: string;
 }
 
-const esClient = new ElasticSearch.Client({
-  hosts: [
-    `https://${Config.esUsername}:${Config.esPassword}@${Config.esHost}:9243`,
-  ],
-});
-
 class SearchController {
   public static async searchByInvno(
     request: express.Request,
     response: express.Response
   ) {
     const invno = request.query.invno;
-    const elasticSearchResponse = (
-      await esClient.search({
-        index: "collection",
-        body: {
-          size: 1,
-          query: {
-            bool: {
-              must: {
-                match: {
-                  invno: invno,
-                },
-              },
-            },
-          },
-          _source: ["id", "invno"],
-        },
-      })
-    ).hits.hits;
+    const elasticSearchResponse = await ElasticSearchService.searchByInvno(
+      invno
+    );
 
     // If there's a valid response from ElasticSearch, then we have a usable ID
     if (elasticSearchResponse.length > 0) {
-      const { id } = elasticSearchResponse[0]._source as ElasticSearchSource;
+      const { id } = elasticSearchResponse[0]._source;
       return response.status(200).json({
         id,
         success: true,
